Added Coral Placer subsystem in Climb subsystem. Additionally changed when command code is called during scheduling.
