+++ Big-Bad-Wolf-2025/src/main/java/frc/robot/commands/FunnelOpenCommand.java
@@ -5,36 +5,36 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.CoralPlacerConstants;
-import frc.robot.subsystems.CoralPlacer;
+import frc.robot.constants.ClimbConstants;
+import frc.robot.subsystems.Climb;
 
 // Creates a new Funnel Open command.
 public class FunnelOpenCommand extends Command 
 {
-  private final CoralPlacer m_CoralPlacer;
+  private final Climb m_Climb;
 
   /**
    * Funnel Open command constructor.
    * @param coralPlacer Coral Placer subsystem.
    */
-  public FunnelOpenCommand(CoralPlacer coralPlacer) 
+  public FunnelOpenCommand(Climb climb) 
   {
-    this.m_CoralPlacer = coralPlacer;
-    addRequirements(m_CoralPlacer);
+    this.m_Climb = climb;
+    addRequirements(m_Climb);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() 
   {
-    // Intentionally Empty.
+    m_Climb.setFunnelWristPosition(ClimbConstants.CORAL_FUNNEL_SEPARATE_POSITION);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() 
   {
-    m_CoralPlacer.setFunnelWristPosition(CoralPlacerConstants.CORAL_FUNNEL_SEPARATE_POSITION);
+    // Intentionally Empty.
   }
 
   // Called once the command ends or is interrupted.
