--- Big-Bad-Wolf-2025/src/main/java/frc/robot/constants/CoralPlacerConstants.java
@@ -1,13 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.constants;
-
-// Initializing constants for Coral Placer.
-public class CoralPlacerConstants 
-{
-    // (REMOVE_BEFORE_COMP) Insert coral funnel constants when determined.
-    public static final double CORAL_FUNNEL_SEPARATE_POSITION = 0;
-    public static final double CORAL_FUNNEL_CLOSED_POSITION = 0;
-}
