+++ Big-Bad-Wolf-2025/src/main/java/frc/robot/subsystems/Climb.java
@@ -17,6 +17,8 @@ public class Climb extends SubsystemBase
 {
   private final TalonFX m_ClimbWristMotorMain = new TalonFX(Hardware.CLIMB_WRIST_MOTOR_MAIN_ID);
   private final TalonFX m_ClimbWristMotorFollower = new TalonFX(Hardware.CLIMB_WRIST_MOTOR_FOLLOWER_ID);
+
+  private final TalonFX m_CoralFunnelMotor;
   
   /**
    * Climb subsystem constructor.
@@ -29,6 +31,8 @@ public class Climb extends SubsystemBase
     // Sets the main climb wrist motor to follow the secondary climb wrist motor.
     Follower climbFollowRequest = new Follower(Hardware.CLIMB_WRIST_MOTOR_MAIN_ID, false); 
     m_ClimbWristMotorFollower.setControl(climbFollowRequest);
+
+    m_CoralFunnelMotor = new TalonFX(Hardware.CORAL_FUNNEL_MOTOR_ID);
   }
 
   /**
@@ -39,6 +43,16 @@ public class Climb extends SubsystemBase
   {
     MotionMagicVoltage climbPositionRequest = new MotionMagicVoltage(climbPosition);
     m_ClimbWristMotorMain.setControl(climbPositionRequest);
+  }
+
+  /**
+   * Sets the wrist position for the funnel wrist motor.
+   * @param funnelPosition Position to set the funnel wrist to.
+   */
+  public void setFunnelWristPosition(double funnelPosition)
+  {
+    MotionMagicVoltage positionRequest = new MotionMagicVoltage(funnelPosition);
+    m_CoralFunnelMotor.setControl(positionRequest);
   }
 
   @Override
